requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
langchain==0.1.16
langchain-openai==0.1.3
mysql-connector-python==8.3.0
pyhumps==3.8.0
python-dotenv==1.0.1